base-files: add user config menu at preinit stage
Add a user menu to configure parameters at preinit stage of boot.
This menu lets the user flip any number of options between 0/1,
the value of the option is saved in /etc/config/user, the values
can be retrieved with "get user.@config[0].<option>"

The default timeout is 4 seconds
The default menu has this options:
1:01_enable_wifi	(to be used to enable the wifi if there is no rfkill)
2:02_disable_firewall	(to be used to disable the firewall)

How to recognize when the device is on user config menu:
* The device status led will blink with pulses, staying most of the
  time off.
* If the netsend option is enabled, udp packets will be sent giving
  details about the available options.
* Using the device console

Example of blinks, including the new pulse style (2s scale)
Slow:                      --------------------____________________
Preinit (Fast):            --__--__--__--__--__--__--__--__--__--__
Failsafe:                  _-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-_-
Preinit_regular:           ----____----____----____----____----____
Pulse (with default time): -___________________-___________________

The pulse style accept 2 parameters to configure it, <ms on> <ms off>,
the default value if no values are given is 50 950, used to signal the
boot menu.

How to choose an option when the device in on the menu:
 * With a console, you can just type the option number.
 * With buttons, you can press any button X times to select option X
   and then wait until the menu ends the time (timeout). At this
   stage if you choose a valid option in the menu (pressed X times...)
   you will see an acknowledge of the valid option with a fast led
   status blink (failsafe) that will last for 1 second, and then the
   user config menu will begin again resetting the time and allowing
   the user to choose more options. When the user presses any button X
   times and X is higher than the available options on the menu, it
   is considered an invalid option and the device will continue
   booting.

Note that if there is too many options, you should consider an appropriate
timeout if you intend to press a button X times... a good choose could
be to consider 1s for each 2 seconds: Timeout = Options /2

While the device is running this menu, the buttons won't work as usual,
this allows the user to press any button regardless of the device, and it
should work even with the reset button. Buttons will work as usual when
the user config menu is over, so if you use the reset button to choice an
option, make sure you don´t press the button out of time... :)

The default user menu can be configured inside the preinit options in
config menu, it can be completely removed, it can be reconfigured or
disabled using a config file once the device is booted.

The options are configured populating files in folder /etc/config-menu.d,
the options will be order by name. When the user choose a valid option,
the script associated with that option in folder /etc/config-menu.d will
be executed with argument 0 or 1 (the current value flipped).

Enable/disable the menu using config file (0 enable, 1 disable):
  uci set user.@config[0].no_user_config=<1|0>
Add new options:
  Copy an script to the folder /etc/config-menu.d
Change the menu timeout (minimum 1):
  uci set user.@config[0].user_config_wait_timeout=<seconds>

This PR is related with but don't require:
PR https://github.com/openwrt/openwrt/pull/2408

Signed-off-by: Manuel Giganto <mgigantoregistros@gmail.com>